+++ firmware/uNode_2/config.h
@@ -9,7 +9,7 @@
 
 #define FW_VERSION_MAJOR 0
 #define FW_VERSION_MINOR 23
-#define FW_VERSION_PATCH 12
+#define FW_VERSION_PATCH 13
 
 #define FW_STRINGIFY_IMPL(value) #value
 #define FW_STRINGIFY(value) FW_STRINGIFY_IMPL(value)
@@ -24,7 +24,7 @@
 #define FW_FLASH_LAYOUT "4M1M"
 #define FW_LITTLEFS_IMAGE_SIZE 0xFA000UL
 #define FW_WEB_ASSET_VERSION FW_VERSION
-#define CONFIG_SCHEMA_VERSION 6
+#define CONFIG_SCHEMA_VERSION 7
 
 #define HEAP_WARNING_FREE_BYTES 12000UL
 #define HEAP_WARNING_MAX_BLOCK_BYTES 6000UL
@@ -281,7 +281,8 @@ enum BusGuardMode {
 
 enum ButtonAction {
   BUTTON_ACTION_DISABLED = 0,
-  BUTTON_ACTION_TOGGLE_LOCATE = 1
+  BUTTON_ACTION_TOGGLE_LOCATE = 1,
+  BUTTON_ACTION_MUTE_LEDS_UNTIL_REBOOT = 2
 };
 
 // -----------------------------------------------------------------------------
@@ -327,7 +328,8 @@ struct Config {
 
   TerminationMode terminationMode;
   BusGuardMode busGuardMode;
-  ButtonAction buttonAction;
+  ButtonAction buttonShortAction;
+  ButtonAction buttonLongAction;
 
   bool legacyArtPollReply;